html/obsolete/requirements-for-implementations/frames/document-frameset-color.html

<!DOCTYPE html>
<title>document frameset color</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>

<script>
test(function() {

var fset = document.createElement("frameset");
document.documentElement.appendChild(fset);

document.fgColor = "blue";
document.bgColor = "blue";
document.linkColor = "blue";
document.vlinkColor = "blue";
document.alinkColor = "blue";

assert_equals(document.fgColor,"");
assert_equals(document.bgColor,"");
assert_equals(document.linkColor,"");
assert_equals(document.vlinkColor,"");
assert_equals(document.alinkColor,"");

document.documentElement.removeChild(fset);

}
//setting the document color attributes with a frameset
);

test(function() {

document.fgColor = "blue";
document.bgColor = "blue";
document.linkColor = "blue";
document.vlinkColor = "blue";
document.alinkColor = "blue";

var fset = document.createElement("frameset");
document.documentElement.appendChild(fset);

assert_equals(document.fgColor,"");
assert_equals(document.bgColor,"");
assert_equals(document.linkColor,"");
assert_equals(document.vlinkColor,"");
assert_equals(document.alinkColor,"");

document.removeChild(fset);

}
//getting the document color attributes with a frameset
);
</script>
